+++ app/src/main/java/com/grusie/sharingmap/designsystem/theme/Color.kt
@@ -1,0 +1,21 @@
+package com.grusie.sharingmap.designsystem.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
+
+val Blue = Color(0xFF458FFF)
+val Red = Color(0xFFF41E00)
+val Black = Color(0xFF080809)
+val White = Color(0xFFFFFFFF)
+
+val Gray100 = Color(0xFFF1F4F7)
+val Gray200 = Color(0xFFE8EAEB)
+val Gray300 = Color(0xFFBABBBC)
+val Gray400 = Color(0xFF9A9C9F)
